1）增加高德Key  2）修改小班行测处的回调数据结构

+++ app/src/main/java/com/appublisher/quizbank/model/business/KnowledgeTreeModel.java
@@ -195,7 +195,7 @@ public class KnowledgeTreeModel {
                             hierarchy.getNotes(),
                             hierarchy.getLevel(),
                             hierarchy.getRight(),
-                            hierarchy.getDuration()));
+                            (int) hierarchy.getDuration()));
         } else {
             return new TreeNode(
                     new TreeItemHolder.TreeItem(
